+++ RacingLeagueHub.Application/GameTeams/Persistence/IGameTeamCommands.cs
@@ -5,6 +5,6 @@ namespace RacingLeagueHub.Application.GameTeams.Persistence;
 public interface IGameTeamCommands
 {
     Task<GameTeamDto> AddAsync(CreateGameTeamDto dto, CancellationToken ct);
-    Task<GameTeamDto?> UpdateAsync(long id, UpdateGameTeamDto dto, CancellationToken ct);
-    Task<bool> DeleteAsync(long id, CancellationToken ct);
+    Task<GameTeamDto?> UpdateAsync(int id, UpdateGameTeamDto dto, CancellationToken ct);
+    Task<bool> DeleteAsync(int id, CancellationToken ct);
 }
