+++ RacingLeagueHub.Application/GameTeams/Persistence/IGameTeamQueries.cs
@@ -5,6 +5,6 @@ namespace RacingLeagueHub.Application.GameTeams.Persistence;
 
 public interface IGameTeamQueries
 {
-    Task<GameTeamDto?> GetByIdAsync(long id, CancellationToken ct);
+    Task<GameTeamDto?> GetByIdAsync(int id, CancellationToken ct);
     Task<PagedResult<GameTeamDto>> GetPagedAsync(int page, int pageSize, CancellationToken ct);
 }
