+++ RacingLeagueHub.Application/LeagueUsers/Dtos/LeagueUserRolesDto.cs
@@ -2,4 +2,4 @@
 
 namespace RacingLeagueHub.Application.LeagueUsers.Dtos;
 
-public record LeagueUserRolesDto (long UserId, long LeagueId, string LeagueSlug, bool IsOwner, bool IsAdmin, bool IsEditor, bool IsSteward);
+public record LeagueUserRolesDto (int UserId, int LeagueId, string LeagueSlug, bool IsOwner, bool IsAdmin, bool IsEditor, bool IsSteward);
