+++ RacingLeagueHub.Application/Resources/IResourceService.cs
@@ -6,10 +6,10 @@ namespace RacingLeagueHub.Application.Resources;
 
 public interface IResourceService
 {
-    Task<ResourceDto?> GetByIdAsync(long id, CancellationToken ct = default);
+    Task<ResourceDto?> GetByIdAsync(int id, CancellationToken ct = default);
     Task<PagedResult<ResourceDto>> GetPagedAsync(int page, int pageSize, CancellationToken ct = default);
     Task<ResourceDto> UploadAsync(FileUploadRequest file, bool? isThumbnail, CancellationToken ct = default);
-    Task ConfirmAsync(long id, CancellationToken ct = default);
-    Task DeleteAsync(long uid, CancellationToken ct = default);
-    Task<string?> GetFileUrlAsync(long id, CancellationToken ct = default);
+    Task ConfirmAsync(int id, CancellationToken ct = default);
+    Task DeleteAsync(int uid, CancellationToken ct = default);
+    Task<string?> GetFileUrlAsync(int id, CancellationToken ct = default);
 }
