+++ RacingLeagueHub.Application/Teams/Persistence/ITeamCommands.cs
@@ -5,6 +5,6 @@ namespace RacingLeagueHub.Application.Teams.Persistence;
 public interface ITeamCommands
 {
     Task<TeamDto> AddAsync(CreateTeamDto dto, CancellationToken ct);
-    Task<TeamDto?> UpdateAsync(long id, UpdateTeamDto dto, CancellationToken ct);
-    Task<bool> DeleteAsync(long id, CancellationToken ct);
+    Task<TeamDto?> UpdateAsync(int id, UpdateTeamDto dto, CancellationToken ct);
+    Task<bool> DeleteAsync(int id, CancellationToken ct);
 }
