+++ RacingLeagueHub.Application/Teams/Persistence/ITeamQueries.cs
@@ -5,6 +5,6 @@ namespace RacingLeagueHub.Application.Teams.Persistence;
 
 public interface ITeamQueries
 {
-    Task<TeamDto?> GetByIdAsync(long id, CancellationToken ct);
+    Task<TeamDto?> GetByIdAsync(int id, CancellationToken ct);
     Task<PagedResult<TeamDto>> GetPagedAsync(int page, int pageSize, CancellationToken ct);
 }
