+++ RacingLeagueHub.Application/TrackLayouts/ITrackLayoutService.cs
@@ -5,8 +5,8 @@ namespace RacingLeagueHub.Application.TrackLayouts;
 
 public interface ITrackLayoutService
 {
-    Task<TrackLayoutDto?> GetByIdAsync(long id, CancellationToken ct);
+    Task<TrackLayoutDto?> GetByIdAsync(int id, CancellationToken ct);
     Task<PagedResult<TrackLayoutDto>> GetPagedAsync(int page, CancellationToken ct);
     Task<TrackLayoutDto> AddAsync(CreateTrackLayoutDto dto, CancellationToken ct);
-    Task<TrackLayoutDto?> UpdateAsync(long id, UpdateTrackLayoutDto dto, CancellationToken ct);
+    Task<TrackLayoutDto?> UpdateAsync(int id, UpdateTrackLayoutDto dto, CancellationToken ct);
 }
