+++ RacingLeagueHub.Domain/Entities/Stewarding/Verdict.cs
@@ -2,9 +2,9 @@
 
 public class Verdict : EntityBase
 {
-    public long IncidentId { get; set; }
+    public int IncidentId { get; set; }
     public string Summary { get; set; }
-    public long Explanation { get; set; }
+    public int Explanation { get; set; }
     public DateTimeOffset CreatedAt { get; set; }
     public short StewardPenaltyType { get; set; }
     public short? PenaltyAmount { get; set; }
