+++ RacingLeagueHub.Domain/Entities/TrackLayoutGame.cs
@@ -4,7 +4,7 @@ namespace RacingLeagueHub.Domain.Entities;
 
 public class TrackLayoutGame
 {
-    public long TrackLayoutId { get; set; }
+    public int TrackLayoutId { get; set; }
     public Game Game { get; set; }
 
     public virtual TrackLayout TrackLayout { get; set; }
